requirements.txt
flask
Pillow
webcolors